Move string functions into own domain
