+++ src/Libraries/2D/Source/Gen/genstr.c
@@ -55,7 +55,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #include "clpcon.h"
 #include "ctxmac.h"
 #include "grdbm.h"
-#include "chr.h"
+#include "string/chr.h"
 
 
 int gen_font_string (grs_font *f, char *s, short x0, short y0)
