+++ src/Libraries/2D/Source/Gen/genustr.c
@@ -52,7 +52,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 #include "bitmap.h"
 #include "grdbm.h"
-#include "chr.h"
+#include "string/chr.h"
 
 
 /* draw a string s in the specified font at (x0,y0).  does not perform
