+++ src/Libraries/2D/Source/string/str.h
