+++ src/Libraries/2D/Source/string/strwid.c
